requirements.txt
fastapi==0.118.0
uvicorn==0.37.0
python-multipart==0.0.20
pillow==11.3.0
requests==2.32.5
pydantic-settings==2.6.1
